Added Progressbar to the bottom of the main window.
The progress bar helps to get a sense of how long searches take and give people something to look at when
waiting for results in the names dictionary to come back.  Also, the search button now no longer disables when
there is no search text because it was morso distracting than helpful.