Initial work to use GtkBuilder files directly.
The official file extension became *.ui from *.xml.  Glade 3.6 can save directly to these files.  The
Interface was reworked to take advantage of GtkAction which is a specialty of GtkBuilder.  There was a fair
amount of code cleanup, and the dictionary names are now totally translated.  Unfortunately, elements are
not being automatically translated with the new format.