Fix for searches of the term "Go"
The evil term Ageru has 25 definitions when I only designed the algorithm for most 20
and with little testing for results with 2 digits.  The safeties have been put in place
and the number has been jumped to 50.  If there are more definitions than that, the last
ones will just be interpreted as one big definition.